Updated for unique filename
Changed the code so it produced unique URL.  Instead of leaving users to change it, changed it in code.  The issue was explained here: https://help.disqus.com/customer/portal/articles/662547
The first time Disqus is loaded on a page, our system creates a new thread ID and associates that thread ID with the page. Unfortunately, if multiple pages are loaded from the same URL that means each page will be associated with the same thread ID.
If page URL or ContentID match, the comments are loaded on a single thread.